# news-aggregatorcomposer/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API Documentation",
        "description": "News Aggregator API Documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:70/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/password/email": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Send password reset link",
                "operationId": "4f2e748c9c87bd82791fbe0fb38ca68c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "example@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset link sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reset link sent!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/password/reset": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Reset user password",
                "operationId": "e6238abb107fefd729cdcff6352305ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "reshmabelman@gmail.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "your-reset-token"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password860"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password860"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password reset successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "User not found or invalid token"
                    }
                }
            }
        },
...
        "/preferences": {
            "get": {
                "tags": [
                    "preference"
                ],
                "summary": "Get user preferences",
                "operationId": "6a58074d66df8ad85bb1edff31fba75b",
                "responses": {
                    "200": {
                        "description": "User preferences retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "preferred_source": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "preferred_category": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "preferred_author": {
                                                    "type": "string",
                                                    "example": "By Michael Gold"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-11-01T13:39:52.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-11-01T14:04:19.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "preference"
                ],
                "summary": "Set user preferences",
                "operationId": "f52a75fb74ca625c4170fa0c0dfe2c15",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "preferred_source": {
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "preferred_category": {
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "preferred_author": {
                                        "type": "string",
                                        "example": "By Michael Gold"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User preferences updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "preferred_source": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "preferred_category": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "preferred_author": {
                                                    "type": "string",
                                                    "example": "By Michael Gold"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-11-01T13:39:52.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-11-01T14:04:19.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentication using Laravel Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}